[etherscan] change plugin structure
