+++ package.json
@@ -98,7 +98,7 @@
     "start": "yarn && node ./scripts/wrapper.js \"webpack serve --config ./scripts/webpack.config.js --env PLUGIN=\"",
     "host": "yarn && node ./scripts/wrapper.js \"webpack serve --config ./scripts/webpack.config.js --mode=production --env PLUGIN=\"",
     "build": "yarn && node ./scripts/wrapper.js \"webpack build --config ./scripts/webpack.config.js --mode=production --env PLUGIN=\"",
-    "test": "eslint ./src && tsc --noEmit && node ./scripts/wrapper.js \"jest \""
+    "test": "eslint ./src && tsc --noEmit && node ./scripts/wrapper.js \"jest --watch \""
   },
   "jest": {
     "preset": "ts-jest",
