+++ src/plugins/etherscan/__tests__/converters/accounts/accounts.test.ts
@@ -1,5 +1,5 @@
 import { AccountType } from '../../../../../types/zenmoney'
-import { convertAccounts } from '../../../converters'
+import { convertAccounts } from '../../../ether/converters'
 
 describe('convertAccounts', () => {
   it.each([
