+++ src/plugins/etherscan/common/types.ts
@@ -1,0 +1,9 @@
+export interface Preferences {
+  apiKey: string
+  account: string
+}
+
+export interface Response {
+  status: string
+  message: string
+}
